set up TICL so that Leiden is active
