+++ DataFormats/HGCalReco/src/classes.h
@@ -6,5 +6,4 @@
 #include "DataFormats/HGCalReco/interface/TICLLayerTile.h"
 #include "DataFormats/HGCalReco/interface/TICLSeedingRegion.h"
 #include "DataFormats/HGCalReco/interface/TICLCandidate.h"
-#include "DataFormats/HGCalReco/interface/TICLGraph.h"
 #include "DataFormats/Common/interface/Wrapper.h"
