+++ RecoHGCal/TICL/plugins/LinkingAlgoFactory.cc
@@ -1,10 +1,5 @@
 #include "FWCore/ParameterSet/interface/ValidatedPluginFactoryMacros.h"
 #include "FWCore/ParameterSet/interface/ValidatedPluginMacros.h"
 #include "LinkingAlgoFactory.h"
-#include "LinkingAlgoByLeiden.h"
 
 EDM_REGISTER_VALIDATED_PLUGINFACTORY(LinkingAlgoFactory, "LinkingAlgoFactory");
-
-DEFINE_EDM_VALIDATED_PLUGIN(LinkingAlgoFactory,
-                            ticl::LinkingAlgoByLeiden,
-                            "LinkingAlgoByLeiden");
